Add files via upload
Initial commit.
Automated my hours of dull work. This script makes a search for lossles audio file image (*.ape, *flac). In a directory it is stored in. And splits into separate tracks using shntool and cuebreakpoints (which need to be installed before using these scripts).
A lot of things to improve:
1. Files should be tagged using *.cue
2. Handling of more than one pair of *.cue and (*.flac or *.ape)
3. Recursion.